Fixes #3: About the zh_cn language problem
There was a wrong `preg_replace`, when questions were added which turned `zh_cn`to `zhcn`.
Since this is nt a valid language code, no questions were added.

After all, this was not a `zh_cn` problem only, but a problem of all languages with a underscore in their language code.

	modified:   lib.php